refactor eventhubs-integration
